perf: conditionally load PostHog only when API key is set
